backend/pyproject.toml
[tool.poetry]
name = "hms-backend"
version = "0.1.0"
description = "HMS API with FastAPI"
authors = ["Admin <admin@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
asyncmy = "^0.2.9"
cryptography = "^42.0.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.1.0"
email-validator = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
